creating an actuall structure
